# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Foot Easy API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/teams": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "List Teams",
        "description": "List teams, optionally filtered by club and season.",
        "operationId": "list_teams_api_v1_teams_get",
        "parameters": [
          {
            "name": "club_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Club Id"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{4}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TeamResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Teams"
        ],
        "summary": "Create Team",
        "description": "Create a club team; the name must be unique within the club and season.",
        "operationId": "create_team_api_v1_teams_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{team_id}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Get Team",
        "description": "Get a team by id.",
        "operationId": "get_team_api_v1_teams__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Teams"
        ],
        "summary": "Partial Update Team",
        "description": "Partially update a team.",
        "operationId": "partial_update_team_api_v1_teams__team_id__patch",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Teams"
        ],
        "summary": "Delete Team",
        "description": "Delete a team with its members, events and invitations.",
        "operationId": "delete_team_api_v1_teams__team_id__delete",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/match-facts": {
      "get": {
        "tags": [
          "Match facts"
        ],
        "summary": "List Match Facts",
        "description": "List a match's goals and cards in chronological order.",
        "operationId": "list_match_facts_api_v1_match_facts_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_MatchFactResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Match facts"
        ],
        "summary": "Create Match Fact",
        "description": "Record a goal (with optional assist) or a card for one of the team's players.",
        "operationId": "create_match_fact_api_v1_match_facts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchFactCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchFactResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/match-facts/{fact_id}": {
      "delete": {
        "tags": [
          "Match facts"
        ],
        "summary": "Delete Match Fact",
        "description": "Delete a match fact.",
        "operationId": "delete_match_fact_api_v1_match_facts__fact_id__delete",
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Fact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/teams/{team_id}": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Get Team Stats",
        "description": "Season record of a team and its players: goals, assists, cards, selections, attendance.",
        "operationId": "get_team_stats_api_v1_stats_teams__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamStats"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "FactKind": {
        "type": "string",
        "enum": [
          "goal",
          "yellow_card",
          "red_card"
        ],
        "title": "FactKind"
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
...
      "InvitedMember": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "shirt_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shirt Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "role",
          "position",
          "shirt_number"
        ],
        "title": "InvitedMember"
      },
...
      "MatchFactCreate": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "kind": {
            "$ref": "#/components/schemas/FactKind"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "assist_member_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assist Member Id"
          },
          "minute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 130.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "kind",
          "member_id"
        ],
        "title": "MatchFactCreate"
      },
      "MatchFactResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "kind": {
            "$ref": "#/components/schemas/FactKind"
          },
          "minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute"
          },
          "member": {
            "$ref": "#/components/schemas/InvitedMember"
          },
          "assist_member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvitedMember"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "kind",
          "minute",
          "member",
          "assist_member"
        ],
        "title": "MatchFactResponse"
      },
...
      "MemberRole": {
        "type": "string",
        "enum": [
          "player",
          "coach",
          "staff"
        ],
        "title": "MemberRole"
      },
...
      "PaginatedResponse_MatchFactResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchFactResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse[MatchFactResponse]"
      },
...
      "PaginatedResponse_TeamResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TeamResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse[TeamResponse]"
      },
      "PlayerPosition": {
        "type": "string",
        "enum": [
          "goalkeeper",
          "defender",
          "midfielder",
          "forward"
        ],
        "title": "PlayerPosition"
      },
      "PlayerStats": {
        "properties": {
          "member": {
            "$ref": "#/components/schemas/InvitedMember"
          },
          "selections": {
            "type": "integer",
            "title": "Selections"
          },
          "goals": {
            "type": "integer",
            "title": "Goals"
          },
          "assists": {
            "type": "integer",
            "title": "Assists"
          },
          "yellow_cards": {
            "type": "integer",
            "title": "Yellow Cards"
          },
          "red_cards": {
            "type": "integer",
            "title": "Red Cards"
          },
          "invited": {
            "type": "integer",
            "title": "Invited"
          },
          "present": {
            "type": "integer",
            "title": "Present"
          },
          "attendance_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attendance Rate"
          }
        },
        "type": "object",
        "required": [
          "member",
          "selections",
          "goals",
          "assists",
          "yellow_cards",
          "red_cards",
          "invited",
          "present",
          "attendance_rate"
        ],
        "title": "PlayerStats"
      },
...
      "TeamCategory": {
        "type": "string",
        "enum": [
          "u7",
          "u9",
          "u11",
          "u13",
          "u15",
          "u17",
          "u19",
          "senior",
          "veteran"
        ],
        "title": "TeamCategory"
      },
      "TeamCreate": {
        "properties": {
          "club_id": {
            "type": "string",
            "format": "uuid",
            "title": "Club Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/TeamCategory"
          },
          "season": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{4}$",
            "title": "Season",
            "examples": [
              "2026-2027"
            ]
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Color",
            "default": "#16a34a",
            "examples": [
              "#16a34a"
            ]
          }
        },
        "type": "object",
        "required": [
          "club_id",
          "name",
          "category",
          "season"
        ],
        "title": "TeamCreate"
      },
      "TeamResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "club_id": {
            "type": "string",
            "format": "uuid",
            "title": "Club Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/TeamCategory"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "club_id",
          "name",
          "category",
          "season",
          "color",
          "created_at"
        ],
        "title": "TeamResponse"
      },
      "TeamStats": {
        "properties": {
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "played": {
            "type": "integer",
            "title": "Played"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "draws": {
            "type": "integer",
            "title": "Draws"
          },
          "losses": {
            "type": "integer",
            "title": "Losses"
          },
          "goals_for": {
            "type": "integer",
            "title": "Goals For"
          },
          "goals_against": {
            "type": "integer",
            "title": "Goals Against"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerStats"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "played",
          "wins",
          "draws",
          "losses",
          "goals_for",
          "goals_against",
          "players"
        ],
        "title": "TeamStats"
      },
      "TeamUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "season": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{4}$",
                "examples": [
                  "2026-2027"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$",
                "examples": [
                  "#16a34a"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "title": "TeamUpdate"
      },
...
    }
  }
}